MonopolyDeluxe/requirements.txt
numpy
scikit-image
scipy
matplotlib
pillow